Decrease the size of DepCleanGradleAction and introduce two new utilities classes for Action class
